Refactor stat card data calculations
