+++ packages/ui/src/components/popcorn/emissions-dashboard/ContractContainer/index.tsx
@@ -1,10 +1,7 @@
-import { CloudIcon } from '@heroicons/react/outline';
-import { getChartData, percentChange } from '@popcorn/utils';
+import { getChartData, getStatCardData } from '@popcorn/utils';
 import React from 'react';
-import { Globe } from 'react-feather';
 import {
   Contract,
-  StatCardData,
   Transaction,
 } from '../../../../interfaces/emissions-dashboard';
 import { BiaxialLineChart } from '../recharts/BiaxialLineChart';
@@ -17,47 +14,6 @@ interface ContractContainerProps {
   endDate: Date;
   contract: Contract;
 }
-
-const getStatCardData = (
-  transactionsCurrentPeriod: Transaction[],
-  transactionsPreviousPeriod: Transaction[],
-): StatCardData[] => {
-  const totalEmissionsCurrentPeriod = Math.round(
-    transactionsCurrentPeriod.reduce((acc, cu) => acc + cu.emissions, 0),
-  );
-  const totalEmissionsPreviousPeriod = Math.round(
-    transactionsPreviousPeriod.reduce((acc, cu) => acc + cu.emissions, 0),
-  );
-  const emissionsChangePercentChange = percentChange(
-    totalEmissionsPreviousPeriod,
-    totalEmissionsCurrentPeriod,
-  );
-
-  const totalTransactionVolCurrentPeriod = transactionsCurrentPeriod.length;
-  const totalTransactionVolPreviousPeriod = transactionsPreviousPeriod.length;
-  const transactionVolPercentChange = percentChange(
-    totalTransactionVolPreviousPeriod,
-    totalTransactionVolCurrentPeriod,
-  );
-  return [
-    {
-      id: 1,
-      name: 'CO2 Emissions (µg)',
-      stat: totalEmissionsCurrentPeriod,
-      icon: CloudIcon,
-      change: `${Math.round(emissionsChangePercentChange)}%`,
-      changeType: emissionsChangePercentChange > 0 ? 'increase' : 'decrease',
-    },
-    {
-      id: 2,
-      name: 'Transactions',
-      stat: totalTransactionVolCurrentPeriod,
-      icon: Globe,
-      change: `${transactionVolPercentChange}%`,
-      changeType: transactionVolPercentChange > 0 ? 'increase' : 'decrease',
-    },
-  ];
-};
 
 export const ContractContainer: React.FC<ContractContainerProps> = ({
   transactionsCurrentPeriod,
@@ -78,6 +34,7 @@ export const ContractContainer: React.FC<ContractContainerProps> = ({
           stats={getStatCardData(
             transactionsCurrentPeriod,
             transactionsPreviousPeriod,
+            false,
           )}
         />
       </div>
