+++ packages/ui/src/components/popcorn/emissions-dashboard/TotalStats/index.tsx
@@ -1,11 +1,6 @@
-import { CloudIcon } from '@heroicons/react/outline';
-import {
-  StatCardData,
-  Transaction,
-} from '@popcorn/ui/interfaces/emissions-dashboard';
-import { getChartData, percentChange } from '@popcorn/utils';
+import { Transaction } from '@popcorn/ui/interfaces/emissions-dashboard';
+import { getChartData, getStatCardData } from '@popcorn/utils';
 import React from 'react';
-import { Globe, Wind } from 'react-feather';
 import { BiaxialLineChart } from '../recharts/BiaxialLineChart';
 import { StatsCards } from '../StatsCards';
 
@@ -15,83 +10,6 @@ interface TotalStatsProps {
   startDate: Date;
   endDate: Date;
 }
-
-const GWEI_TO_ETH = Math.pow(10, 9);
-
-const getStatCardData = (
-  transactionsCurrentPeriod: Transaction[],
-  transactionsPreviousPeriod: Transaction[],
-): StatCardData[] => {
-  const totalEmissionsCurrentPeriod = Math.round(
-    transactionsCurrentPeriod.reduce((acc, cu) => acc + cu.emissions, 0),
-  );
-  const totalEmissionsPreviousPeriod = Math.round(
-    transactionsPreviousPeriod.reduce((acc, cu) => acc + cu.emissions, 0),
-  );
-  const emissionsChangePercentChange = percentChange(
-    totalEmissionsPreviousPeriod,
-    totalEmissionsCurrentPeriod,
-  );
-
-  const totalTransactionVolCurrentPeriod = transactionsCurrentPeriod.length;
-  const totalTransactionVolPreviousPeriod = transactionsPreviousPeriod.length;
-  const transactionVolPercentChange = percentChange(
-    totalTransactionVolPreviousPeriod,
-    totalTransactionVolCurrentPeriod,
-  );
-
-  const averageGasPriceCurrentPeriod =
-    transactionsCurrentPeriod.length === 0
-      ? 0
-      : Math.round(
-          transactionsPreviousPeriod.reduce(
-            (acc, cu) => acc + Number(cu.gasPrice),
-            0,
-          ) /
-            (GWEI_TO_ETH * transactionsCurrentPeriod.length),
-        );
-  const averageGasPricePreviousPeriod =
-    transactionsPreviousPeriod.length === 0
-      ? 0
-      : Math.round(
-          transactionsPreviousPeriod.reduce(
-            (acc, cu) => acc + Number(cu.gasPrice),
-            0,
-          ) /
-            (GWEI_TO_ETH * transactionsPreviousPeriod.length),
-        );
-
-  const gasPricePercentChange = percentChange(
-    averageGasPricePreviousPeriod,
-    averageGasPriceCurrentPeriod,
-  );
-  return [
-    {
-      id: 1,
-      name: 'CO2 Emissions (µg)',
-      stat: totalEmissionsCurrentPeriod,
-      icon: CloudIcon,
-      change: `${Math.round(emissionsChangePercentChange)}%`,
-      changeType: emissionsChangePercentChange > 0 ? 'increase' : 'decrease',
-    },
-    {
-      id: 2,
-      name: 'Transactions',
-      stat: totalTransactionVolCurrentPeriod,
-      icon: Globe,
-      change: `${transactionVolPercentChange}%`,
-      changeType: transactionVolPercentChange > 0 ? 'increase' : 'decrease',
-    },
-    {
-      id: 3,
-      name: 'Average Gas Price',
-      stat: averageGasPriceCurrentPeriod,
-      icon: Wind,
-      change: `${gasPricePercentChange}%`,
-      changeType: gasPricePercentChange > 0 ? 'increase' : 'decrease',
-    },
-  ];
-};
 
 export const TotalStats: React.FC<TotalStatsProps> = ({
   transactionsCurrentPeriod,
@@ -118,6 +36,7 @@ export const TotalStats: React.FC<TotalStatsProps> = ({
           stats={getStatCardData(
             transactionsCurrentPeriod,
             transactionsPreviousPeriod,
+            true,
           )}
         />
       </div>
